fixed bug
